Switched to OpenVZ, made everything work... I am so sorry to anyone who ever looks at this commit
The switch to OpenVZ was required because of the issues Docker was
causing with networking and binding to core.
There is now complete automation of the data collection. You just have
to input the tests to run in OpenVZ/runscript/testlist.txt and run
autoplot.py. Additionally, the majority of the scripts in runscript have
been rewritten in python.

Over the next couple of weeks I wil try to make the code readable and
actually work on any setup other than my own without major modification.